Add a ImageBase class which Image and Texture inherit from.
This is used so that Images and Textures can have different
implementations of constructDescriptorImageInfo and
getPrimaryImageUsageFlags but the memory allocation etc can still be
common.

This fixes a number of validation errors about mismatching image usage
bits etc.

Also fix a memory leak of the Sampler created when the texture is
created.